Use `evalMapChunk` to preserve underlying stream chunks in consumers
This applies to both the gRPC and HTTP consumers. It could lead to
performance improvements in applications which are sensitive to the
chunksize.

There is a subtle user-facing change as a result: if any message in a
chunk fails to decode the `errorHandler` is invoked _before_ any
successfully decoded messages in the chunk are emitted downstream; even
if the message which failed to decode comes after the successful ones.

This change is probably not a problem though...